reorganize modal to live in queue dir
